% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{get_as_mapping}
\alias{get_as_mapping}
\title{Sanity check the 'as' argument
Ensures all values are valid and that multiple values are named.}
\usage{
get_as_mapping(as, choices)
}
\description{
Sanity check the 'as' argument
Ensures all values are valid and that multiple values are named.
}
\keyword{internal}
