% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{get_best_dtype}
\alias{get_best_dtype}
\title{Helper to find the smallest fitting data type for numeric data}
\usage{
get_best_dtype(data, dtype)
}
\arguments{
\item{data}{The numeric or integer vector.}
}
\value{
A string representing the best HDF5 data type.
}
\description{
Helper to find the smallest fitting data type for numeric data
}
\keyword{internal}
