% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{h5_delete}
\alias{h5_delete}
\title{Delete an HDF5 Object}
\usage{
h5_delete(file, name)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{The full path of the object to delete (e.g., \code{"/data/dset"} or \code{"/groups/g1"}).}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effects.
}
\description{
Deletes an object (dataset or group) from an HDF5 file.
If the object is a group, all objects contained within it will be deleted
recursively.
}
\examples{
file <- tempfile(fileext = ".h5")
h5_write(file, "/g1/d1", 1:10)
h5_write(file, "d2", 1:5)
print(h5_ls(file, recursive = TRUE))

# Delete a dataset
h5_delete(file, "d2")
print(h5_ls(file, recursive = TRUE))

# Delete a group (and its contents)
h5_delete(file, "g1")
print(h5_ls(file, recursive = TRUE))
unlink(file)
}
\seealso{
\code{\link[=h5_delete_attr]{h5_delete_attr()}}
}
