% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{h5_delete_group}
\alias{h5_delete_group}
\title{Delete an HDF5 Group}
\usage{
h5_delete_group(file, name)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{The full path of the group to delete (e.g., "/g1/g2").}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effects.
}
\description{
Deletes a group and all objects contained within it.
This function will not delete a dataset.
}
\examples{
file <- tempfile(fileext = ".h5")
h5_write(file, "/g1/g2/dset", 1:10)
print(h5_ls(file, recursive = TRUE)) # "g1" "g1/g2" "g1/g2/dset"

h5_delete_group(file, "/g1")
print(h5_ls(file, recursive = TRUE)) # character(0)
unlink(file)
}
\seealso{
\code{\link[=h5_delete_dataset]{h5_delete_dataset()}}, \code{\link[=h5_delete_attr]{h5_delete_attr()}}
}
