% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5lite.r
\name{h5_dim_attr}
\alias{h5_dim_attr}
\title{Get HDF5 Attribute Dimensions}
\usage{
h5_dim_attr(file, name, attribute)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{Name of the object attached to.}

\item{attribute}{Name of the attribute.}
}
\description{
Returns the dimensions of an attribute.
}
\examples{
file <- tempfile(fileext = ".h5")
h5_write(file, "data", 1)

h5_write_attr(file, "data", "vec_attr", 1:10)
h5_dim_attr(file, "data", "vec_attr") # 10

unlink(file)
}
