% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.r
\name{h5_exists_attr}
\alias{h5_exists_attr}
\title{Check if an HDF5 Attribute Exists}
\usage{
h5_exists_attr(file, name, attribute)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{The path to the object (dataset or group).}

\item{attribute}{The name of the attribute to check.}
}
\value{
A logical value: \code{TRUE} if the attribute exists, \code{FALSE} otherwise.
}
\description{
Checks for the existence of an attribute on an HDF5 object without
raising an error if it does not exist.
}
\examples{
file <- tempfile(fileext = ".h5")
h5_write(file, "my_data", 1)
h5_write_attr(file, "my_data", "units", "meters")

h5_exists_attr(file, "my_data", "units") # TRUE
h5_exists_attr(file, "my_data", "nonexistent_attr") # FALSE

unlink(file)
}
\seealso{
\code{\link[=h5_exists]{h5_exists()}}
}
