% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.r
\name{h5_typeof_attr}
\alias{h5_typeof_attr}
\title{Get HDF5 Attribute Type}
\usage{
h5_typeof_attr(file, name, attribute)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{Name of the object attached to.}

\item{attribute}{Name of the attribute.}
}
\value{
A character string representing the HDF5 storage type.
}
\description{
Returns the low-level HDF5 storage type of an attribute.
}
\examples{
file <- tempfile(fileext = ".h5")
h5_write(file, "data", 1)

h5_write_attr(file, "data", "meta", "info", dims = NULL)
h5_typeof_attr(file, "data", "meta") # "string"

unlink(file)
}
\seealso{
\code{\link[=h5_typeof]{h5_typeof()}}, \code{\link[=h5_class_attr]{h5_class_attr()}}, \code{\link[=h5_exists_attr]{h5_exists_attr()}}
}
