% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.r
\name{print.inspect}
\alias{print.inspect}
\title{Print method for HDF5 inspect objects}
\usage{
\method{print}{inspect}(x, ...)
}
\arguments{
\item{x}{An object of class \code{inspect}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for HDF5 inspect objects
}
\keyword{internal}
