% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{resolve_h5_type}
\alias{resolve_h5_type}
\title{Resolves the HDF5 type based on the 'as' map and data properties}
\usage{
resolve_h5_type(data, name, as_map)
}
\arguments{
\item{data}{The R object.}

\item{name}{The name of the object (column name, dataset name, or attribute name).}

\item{as_map}{The processed 'as' argument (named vector).}
}
\description{
Resolves the HDF5 type based on the 'as' map and data properties
}
\keyword{internal}
