% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.r
\name{validate_strings}
\alias{validate_strings}
\title{Validates string inputs}
\usage{
validate_strings(file, name, attr = NULL, must_exist = FALSE)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{The full path of the object (group or dataset).}

\item{attr}{The name of an attribute to check. If provided, the length of the attribute is returned.}

\item{must_exist}{Logical. If \code{TRUE}, the function will stop if the object does not exist.}
}
\value{
The expanded path to the file.
}
\description{
Validates string inputs
}
\keyword{internal}
