% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write_data}
\alias{write_data}
\title{Write a single dataset or attribute}
\usage{
write_data(data, file, name, attr, as_map, compress = FALSE, dry = FALSE)
}
\description{
Write a single dataset or attribute
}
\keyword{internal}
