% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write_group}
\alias{write_group}
\title{Recursively write a list as a group}
\usage{
write_group(data, file, name, as_map, compress = TRUE, dry = FALSE)
}
\description{
Recursively write a list as a group
}
\keyword{internal}
