% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{h5_delete}
\alias{h5_delete}
\title{Delete an HDF5 Dataset}
\usage{
h5_delete(file, name)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{The full path of the dataset to delete.}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effects.
}
\description{
Deletes a dataset from an HDF5 file. This function will not delete a group.
}
\examples{
file <- tempfile(fileext = ".h5")
h5_write(file, "dset1", 1:10)
h5_write(file, "dset2", 1:5)
print(h5_ls(file))

h5_delete(file, "dset1")
print(h5_ls(file))
unlink(file)
}
\seealso{
\code{\link[=h5_delete_attr]{h5_delete_attr()}}, \code{\link[=h5_delete_group]{h5_delete_group()}}
}
