% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5lite.r
\name{h5_delete}
\alias{h5_delete}
\title{Delete an HDF5 Dataset}
\usage{
h5_delete(file, name)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{The full path of the dataset to delete.}
}
\description{
Deletes a dataset from an HDF5 file.
}
\examples{
file <- tempfile(fileext = ".h5")
h5_write(file, "dset1", 1:10)
h5_write(file, "dset2", 1:5)
h5_ls(file)

h5_delete(file, "dset1")
h5_ls(file)
unlink(file)
}
