% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{h5_delete_attr}
\alias{h5_delete_attr}
\title{Delete an HDF5 Attribute}
\usage{
h5_delete_attr(file, name, attribute)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{The path to the object (dataset or group).}

\item{attribute}{The name of the attribute to delete.}
}
\description{
Deletes an attribute from an object.
}
\examples{
file <- tempfile(fileext = ".h5")
h5_write(file, "data", 1)
h5_write_attr(file, "data", "attr1", 123, dims = NULL)
h5_ls_attr(file, "data") # "attr1"

h5_delete_attr(file, "data", "attr1")
h5_ls_attr(file, "data") # character(0)
unlink(file)
}
