% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.r
\name{h5_ls_attr}
\alias{h5_ls_attr}
\title{List HDF5 Attributes}
\usage{
h5_ls_attr(file, name)
}
\arguments{
\item{file}{Path to the HDF5 file.}

\item{name}{The path to the object (dataset or group) to query.
Use "/" for the file's root attributes.}
}
\value{
A character vector of attribute names.
}
\description{
Lists the names of attributes attached to a specific HDF5 object.
}
\seealso{
\code{\link[=h5_ls]{h5_ls()}}
}
