% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5lite-package.r
\docType{package}
\name{h5lite-package}
\alias{h5lite}
\alias{h5lite-package}
\title{h5lite: Simplified HDF5 Interface}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a user-friendly interface for reading and writing 'HDF5' datasets and attributes. Uses the 'HDF5' library developed by 'The HDF Group' \url{https://www.hdfgroup.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cmmr.github.io/h5lite/}
}

}
\author{
\strong{Maintainer}: Daniel P. Smith \email{dansmith01@gmail.com}

}
\keyword{internal}
